feat: jwt filter to have reqUser on rentals routes

+++ src/main/java/oc/chatopbackend/model/UserModel.java
@@ -8,13 +8,13 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class UserModel {
-    private int id;
+    private Long id;
     private String email;
     private String name;
     private String created_at;
     private String updated_at;
 
-    public UserModel(int id, String email, String name, LocalDateTime created_at, LocalDateTime updated_at) {
+    public UserModel(Long id, String email, String name, LocalDateTime created_at, LocalDateTime updated_at) {
         this.id = id;
         this.email = email;
         this.name = name;
